Added Filter Contour, LFO, Modulation, MOD VCA, OSC1 (WIP)
Keyboard Freq/MIDI required to be passed to the Oscillators in addition to the other data already.
Possibly able to make OSC1 and OSC2 identical.
As of this moment it is unknown if many of the "added" features work as they are not currently hooked up to anything that can be used to test them.