Added ActiveSoundStruct and Reworked State Calculation
Added `UWeatherState::GetLinearVolumeLevel(float State)` too.
Next work is adding `FootstepManagerComponent`s to 'floors' to represent the base sounds - will use temporary sounds, maybe for the release of this too. Simply saying 'Grass' combined with 'Rain' as the audio files.

Likely will need to dynamically create and destroy `AmbientSoundComponent`s (or create them as needed and leave them 'active' in a pool, maybe with a timeout) so that each sound can be played, per foot.